Refactor chat list
