+++ src/components/chats/ChatList/ChatList.tsx
@@ -1,22 +1,14 @@
 import useInfiniteScrollData from '@/components/chats/ChatList/hooks/useInfiniteScrollData'
 import Container from '@/components/Container'
-import MessageModal from '@/components/modals/MessageModal'
 import ScrollableContainer from '@/components/ScrollableContainer'
 import { CHAT_PER_PAGE } from '@/constants/chat'
 import useFilterBlockedMessageIds from '@/hooks/useFilterBlockedMessageIds'
-import useLastReadMessageId from '@/hooks/useLastReadMessageId'
-import usePrevious from '@/hooks/usePrevious'
-import useWrapInRef from '@/hooks/useWrapInRef'
 import { useConfigContext } from '@/providers/ConfigProvider'
 import { getPostQuery } from '@/services/api/query'
 import { useCommentIdsByPostId } from '@/services/subsocial/commentIds'
-import { useMessageData } from '@/stores/message'
 import { useMyAccount } from '@/stores/my-account'
 import { cx } from '@/utils/class-names'
-import { getChatPageLink, getUrlQuery } from '@/utils/links'
-import { validateNumber } from '@/utils/strings'
-import { replaceUrl, sendMessageToParentWindow } from '@/utils/window'
-import { useRouter } from 'next/router'
+import { sendMessageToParentWindow } from '@/utils/window'
 import {
   ComponentProps,
   useEffect,
@@ -26,16 +18,14 @@ import {
   useState,
 } from 'react'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import urlJoin from 'url-join'
 import CenterChatNotice from './CenterChatNotice'
-import MemoizedChatItemWrapper from './ChatItemWrapper'
+import MemoizedChatItemWrapper from './ChatItemWithMenu'
+import ChatListSupportingContent from './ChatListSupportingContent'
 import ChatLoading from './ChatLoading'
 import ChatTopNotice from './ChatTopNotice'
 import useFocusedLastMessageId from './hooks/useFocusedLastMessageId'
-import useIsAtBottom from './hooks/useIsAtBottom'
 import useLoadMoreIfNoScroll from './hooks/useLoadMoreIfNoScroll'
 import useScrollToMessage from './hooks/useScrollToMessage'
-import { NewMessageNotice } from './NewMessageNotice'
 import PinnedMessage from './PinnedMessage'
 
 export type ChatListProps = ComponentProps<'div'> & {
@@ -64,14 +54,8 @@ function ChatListContent({
   newMessageNoticeClassName,
   ...props
 }: ChatListProps) {
-  const router = useRouter()
   const { enableBackButton } = useConfigContext()
-
-  const [initialNewMessageCount, setInitialNewMessageCount] = useState(0)
   const lastReadId = useFocusedLastMessageId(chatId)
-  const [recipient, setRecipient] = useState('')
-  const [messageModalMsgId, setMessageModalMsgId] = useState('')
-  const prevMessageModalMsgId = usePrevious(messageModalMsgId)
 
   const scrollableContainerId = useId()
 
@@ -131,65 +115,6 @@ function ChatListContent({
       unpause: () => setIsPausedLoadMore(false),
     }
   )
-
-  // TODO: refactor this by putting the url query getter logic to ChatPage
-  const hasScrolledToMessageRef = useRef(false)
-  const filteredMessageIdsRef = useWrapInRef(filteredMessageIds)
-  useEffect(() => {
-    if (hasScrolledToMessageRef.current) return
-    hasScrolledToMessageRef.current = true
-
-    const messageId = getUrlQuery('messageId')
-    const recipient = getUrlQuery('targetAcc')
-    const isMessageIdsFetched = rawMessageIds !== undefined
-
-    if (!isMessageIdsFetched) return
-
-    if (!messageId || !validateNumber(messageId)) {
-      if (lastReadId) {
-        scrollToMessage(lastReadId ?? '', {
-          shouldHighlight: false,
-          smooth: false,
-        }).then(() => {
-          const lastReadIdIndex = filteredMessageIdsRef.current.findIndex(
-            (id) => id === lastReadId
-          )
-          const newMessageCount =
-            lastReadIdIndex === -1
-              ? 0
-              : filteredMessageIdsRef.current.length - lastReadIdIndex - 1
-
-          sendMessageToParentWindow(
-            'unread',
-            (filteredMessageIds.length ?? 0).toString()
-          )
-
-          setInitialNewMessageCount(newMessageCount)
-        })
-      }
-      return
-    }
-
-    setMessageModalMsgId(messageId)
-    setRecipient(recipient)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [rawMessageIds, filteredMessageIdsRef, hasScrolledToMessageRef])
-
-  const { setLastReadMessageId } = useLastReadMessageId(chatId)
-  useEffect(() => {
-    const lastId = rawMessageIds?.[rawMessageIds.length - 1]
-    if (!lastId) return
-    setLastReadMessageId(lastId)
-  }, [setLastReadMessageId, rawMessageIds])
-
-  useEffect(() => {
-    if (messageModalMsgId) {
-      replaceUrl(urlJoin(getChatPageLink(router), `/${messageModalMsgId}`))
-    } else if (prevMessageModalMsgId && !messageModalMsgId) {
-      replaceUrl(getChatPageLink(router))
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [prevMessageModalMsgId, messageModalMsgId])
 
   const myAddress = useMyAccount((state) => state.address)
   const { data: chat } = getPostQuery.useQuery(chatId)
@@ -268,54 +193,18 @@ function ChatListContent({
           </InfiniteScroll>
         </Component>
       </ScrollableContainer>
-      <MessageModal
+
+      <ChatListSupportingContent
+        chatId={chatId}
         hubId={hubId}
-        isOpen={!!messageModalMsgId}
-        closeModal={() => setMessageModalMsgId('')}
-        messageId={messageModalMsgId}
-        scrollToMessage={scrollToMessage}
-        recipient={recipient}
-      />
-      <Component>
-        <div className='relative'>
-          <NewMessageNotice
-            key={initialNewMessageCount}
-            className={cx(
-              'absolute bottom-2 right-3',
-              newMessageNoticeClassName
-            )}
-            initialNewMessageCount={initialNewMessageCount}
-            messageIds={messageIds}
-            scrollContainerRef={scrollContainerRef}
-          />
-        </div>
-      </Component>
-      <ScrollToBottom
-        loadedMessageLength={loadedMessageQueries.length}
+        filteredMessageIds={filteredMessageIds}
+        loadedMessageQueries={loadedMessageQueries}
+        rawMessageIds={rawMessageIds}
         scrollContainerRef={scrollContainerRef}
+        scrollToMessage={scrollToMessage}
+        asContainer={asContainer}
+        newMessageNoticeClassName={newMessageNoticeClassName}
       />
     </div>
   )
-}
-
-function ScrollToBottom({
-  scrollContainerRef,
-  loadedMessageLength,
-}: {
-  scrollContainerRef: React.RefObject<HTMLDivElement>
-  loadedMessageLength: number
-}) {
-  const isAtBottom = useIsAtBottom(scrollContainerRef, 100)
-  const replyTo = useMessageData((state) => state.replyTo)
-
-  const isAtBottomRef = useWrapInRef(isAtBottom)
-  useEffect(() => {
-    if (!isAtBottomRef.current) return
-    scrollContainerRef.current?.scrollTo({
-      top: scrollContainerRef.current?.scrollHeight,
-      behavior: 'auto',
-    })
-  }, [loadedMessageLength, isAtBottomRef, scrollContainerRef, replyTo])
-
-  return null
 }
